LPS-195477 Inline class EntityModelListenerRegistry in class AnalyticsConfigurationRegistryImpl
